[APD-7] * Get the low hanging fruit
