+++ cases/serializers.py
@@ -5,14 +5,11 @@ from django.contrib.auth import get_user_model
 from django.utils import timezone
 from rest_framework import serializers
 from rest_framework.settings import api_settings
-from rest_framework.fields import empty
 from cases.models import (Officer, Incident,
                           Offense, IncidentInvolvedParty,
                           IncidentFile, Address, State,
                           City)
-from cases.utils import (convert_date_string_to_object,
-                         handle_incident_foreign_keys_for_creation,
-                         parse_and_create_address)
+from cases.utils import convert_date_string_to_object
 User = get_user_model()
 logger = logging.getLogger('cases')
 
@@ -206,8 +203,6 @@ class IncidentInvolvedPartySerializer(serializers.ModelSerializer):
     def update(self, instance, validated_data):
         logger.debug(f"Validated data: {validated_data}")
         logger.debug(f"instance.ID: {instance.id}")
-        # home_address = validated_data.pop("home_address", None)
-        # employer_address = validated_data.pop("employer_address", None)
 
         for attr in validated_data:
             setattr(instance, attr, validated_data[attr])
@@ -242,7 +237,7 @@ class IncidentFileSerializer(serializers.ModelSerializer):
         read_only_fields = ("id",)
 
 
-def jwt_response_payload_handler(token, user=None, request=None):
+def jwt_response_payload_handler(token, user=None):
     officer = user.officer_set.first()
     officer_data = OfficerSerializer(officer).data
     return {
